% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represent.R
\name{represent}
\alias{represent}
\title{Create a representative dataset post record-linkage.}
\usage{
represent(data, linkage, method, parallel = TRUE, cores = NULL, ...)
}
\arguments{
\item{data}{A data frame of records to be represented.}

\item{linkage}{A numeric vector indication the cluster ids post-record linkage for each record in \code{data}.}

\item{method}{Which method to use for representation. Valid options include "proto_minimax", "proto_random",
and "composite".}

\item{parallel}{Logical flag if to use parallel computation or not (via \code{foreach}).}

\item{cores}{If specified, the number of cores to use with \code{foreach}.}

\item{...}{Additional parameters sent to cluster representation function. See
\link[=clust_proto_minimax]{prototyping} or \link[=clust_composite]{composite} methods.}
}
\description{
Create a representative dataset post record-linkage.
}
\examples{

data("rl_reg1")
rep_dat_random <- represent(rl_reg1, identity.rl_reg1, "proto_random", id = FALSE)
head(rep_dat_random)

}
