% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representr.R
\docType{package}
\name{representr}
\alias{representr}
\alias{representr-package}
\title{representr: A package for creating representative records post-record linkage.}
\description{
The representr package provides two types of representative record generation:
prototyping and composite.
}
\section{Prototyping}{


write about prototyping
}

\section{Composite record generation}{


write about composite
}

