% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{within_category_compare_cpp}
\alias{within_category_compare_cpp}
\title{within_category_compare_cpp
Inner column type record distance function}
\usage{
within_category_compare_cpp(
  ties,
  not_cluster,
  col_type,
  weights,
  orders,
  distance
)
}
\arguments{
\item{ties}{A data frame of the records that are tied}

\item{not_cluster}{A data frame of the records outside the cluster}

\item{col_type}{A vector encoding the column type for each column in the dataset. Can take values
in "categorical", "ordinal", "string", or "numeric"}

\item{weights}{A vector of weights for each column for making some column distances more important. Must sum to 1.
Defaults to equal weights.}

\item{orders}{A named list containing the order of the levels in each ordinal column. Defaults to NULL,
which corresponds to no ordinal variables.}

\item{distance}{function that does nothing right now, but must be supplied to not break other code.}
}
\description{
within_category_compare_cpp
Inner column type record distance function
}
